Split termination test into separate unwind and timeout tests
