+++ tests/script-based-pre/cargo_autoharness_termination_timeout/config.yml
@@ -1,0 +1,4 @@
+# Copyright Kani Contributors
+# SPDX-License-Identifier: Apache-2.0 OR MIT
+script: termination_timeout.sh
+expected: termination_timeout.expected
