+++ tests/script-based-pre/cargo_autoharness_termination_unwind/Cargo.toml
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0 OR MIT
 
 [package]
-name = "cargo_autoharness_termination"
+name = "cargo_autoharness_termination_unwind"
 version = "0.1.0"
 edition = "2024"
