+++ tests/script-based-pre/cargo_autoharness_termination_unwind/config.yml
@@ -1,4 +1,4 @@
 # Copyright Kani Contributors
 # SPDX-License-Identifier: Apache-2.0 OR MIT
-script: termination.sh
-expected: termination.expected
+script: termination_unwind.sh
+expected: termination_unwind.expected
